tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/get-navigatorlanguage-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>NavigatorLanguage: navigator.language returns the user's preferred language</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#navigatorlanguage">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h2>Precondition</h2>
<p>The user agent's preferred language is set as English (en).</p>
<div id="log"></div>
<script>
  test(function() {
    assert_equals(navigator.language, "en");
  });
</script>
